+++ src/main/java/org/ject/support/admin/apply/dto/SubmittedApplyResponse.java
@@ -1,8 +1,9 @@
-package org.ject.support.domain.admin.dto;
+package org.ject.support.admin.apply.dto;
 
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import org.ject.support.domain.admin.dto.ApplicationFormResponse;
 import org.ject.support.domain.apply.domain.ApplicationForm;
 import org.ject.support.domain.apply.domain.Apply;
 import org.ject.support.domain.apply.dto.ApplyPortfolioDto;
